More fully integrate turboframes
