+++ app/helpers/view_components/site_footer.rb
@@ -2,13 +2,10 @@ module ViewComponents
   class SiteFooter < ViewComponent
     extend Mandate::Memoize
 
-    delegate :render_site_header?,
-      :namespace_name, :controller_name,
+    delegate :namespace_name, :controller_name,
       to: :view_context
 
     def to_s
-      return unless render_site_header?
-
       digests = %w[external shared].map do |file|
         Digest::SHA1.hexdigest(File.read(Rails.root.join('app', 'views', 'components', 'footer', "#{file}.html.haml")))
       end
