+++ app/javascript/packs/core.tsx
@@ -9,15 +9,3 @@ import { navigator } from '@hotwired/turbo'
 document.addEventListener('DOMContentLoaded', function () {
   ;(window as any).DOMLoaded = true
 })
-
-document.addEventListener('turbo:before-fetch-request', (e) => {
-  const frame = (e as any).detail.fetchOptions.headers['Turbo-Frame']
-  if (!frame) {
-    return
-  }
-
-  const url = (e as any).detail.url
-  if (url != null) {
-    navigator.history.push(new URL(url))
-  }
-})
