+++ app/views/layouts/application.html.haml
@@ -5,7 +5,7 @@
     -# - unless user_signed_in?
     %meta{ name: "viewport", content: "width=device-width, initial-scale=1" }
 
-    %title Exercism
+    %title= page_title
     = csrf_meta_tags
     = csp_meta_tag
     = action_cable_meta_tag
